moved api view to api_views.py
