+++ django_app/redbox_app/urls.py
@@ -5,12 +5,11 @@ from django.contrib import admin
 from django.urls import include, path
 from django.views.generic.base import RedirectView
 from magic_link import urls as magic_link_urls
-from redbox_core.views.chat_views import ChatViewSet
 from rest_framework import routers
 
 from redbox_app.redbox_core import views
 from redbox_app.redbox_core.views import api_views
-from redbox_app.redbox_core.views.api_views import rate_chat_message
+from redbox_app.redbox_core.views.api_views import ChatViewSet, rate_chat_message
 
 admin.site = AdminSitePlus()
 admin.autodiscover()
